+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Fixed
 
 - `sandstone/Scroller` with `editable` prop to rearrange items properly when moving pointer very fast
+- `sandstone/Scroller` with `editable` prop to complete editing when focus left by 5-way key in pointer mode
 
 ## [2.7.12] - 2023-10-23
 
@@ -41,7 +42,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Fixed
 
 - `sandstone/Panels.Header` to not show `slotAfter` in incorrect position at first rendering when `centered` is given
-- `sandstone/Scroller` to read out properly when `editable` is given 
+- `sandstone/Scroller` to read out properly when `editable` is given
 
 ## [2.7.8] - 2023-08-31